mycdkproject/requirements.txt
aws-cdk-lib==2.33.0
constructs>=10.0.0,<11.0.0
aws-cdk.core
aws-cdk.aws-ssm
aws-cdk.aws-ec2
